+++ fabric/src/main/java/rbasamoyai/createbigcannons/fabric/mixin/CannonMountPointMixin.java
@@ -9,10 +9,12 @@ import net.fabricmc.fabric.api.transfer.v1.transaction.TransactionContext;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import rbasamoyai.createbigcannons.cannon_control.cannon_mount.CannonMountBlockEntity;
 import rbasamoyai.createbigcannons.cannon_control.contraption.AbstractMountedCannonContraption;
 import rbasamoyai.createbigcannons.cannon_control.contraption.PitchOrientedContraptionEntity;
+import rbasamoyai.createbigcannons.cannon_control.fixed_cannon_mount.FixedCannonMountBlockEntity;
 import rbasamoyai.createbigcannons.cannons.big_cannons.breeches.quickfiring_breech.CannonMountPoint;
 
 @Mixin(CannonMountPoint.class)
@@ -27,9 +29,17 @@ public abstract class CannonMountPointMixin extends ArmInteractionPoint {
 		CannonMountPoint self = (CannonMountPoint) (Object) this;
 		boolean simulate = !stack.getOrCreateTag().contains("DontSimulate");
 		if (!simulate) stack.getTag().remove("DontSimulate");
-		if (!(this.getLevel().getBlockEntity(this.pos) instanceof CannonMountBlockEntity mount)) return stack;
-		PitchOrientedContraptionEntity poce = mount.getContraption();
-		if (poce == null || !(poce.getContraption() instanceof AbstractMountedCannonContraption cannon)) return stack;
+		BlockEntity be = this.getLevel().getBlockEntity(this.pos);
+		PitchOrientedContraptionEntity poce;
+		if (be instanceof CannonMountBlockEntity mount) {
+			poce = mount.getContraption();
+		} else if (be instanceof FixedCannonMountBlockEntity mount) {
+			poce = mount.getContraption();
+		} else {
+			return stack;
+		}
+		if (poce == null || !(poce.getContraption() instanceof AbstractMountedCannonContraption cannon))
+			return stack;
 		return self.getInsertedResultAndDoSomething(stack, simulate, cannon, poce);
 	}