+++ forge/src/main/java/rbasamoyai/createbigcannons/forge/mixin/CannonMountPointMixin.java
@@ -8,10 +8,12 @@ import com.simibubi.create.content.kinetics.mechanicalArm.ArmInteractionPointTyp
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import rbasamoyai.createbigcannons.cannon_control.cannon_mount.CannonMountBlockEntity;
 import rbasamoyai.createbigcannons.cannon_control.contraption.AbstractMountedCannonContraption;
 import rbasamoyai.createbigcannons.cannon_control.contraption.PitchOrientedContraptionEntity;
+import rbasamoyai.createbigcannons.cannon_control.fixed_cannon_mount.FixedCannonMountBlockEntity;
 import rbasamoyai.createbigcannons.cannons.big_cannons.breeches.quickfiring_breech.CannonMountPoint;
 
 @Mixin(CannonMountPoint.class)
@@ -24,9 +26,17 @@ public abstract class CannonMountPointMixin extends ArmInteractionPoint {
 	@Override
 	public ItemStack insert(ItemStack stack, boolean simulate) {
 		CannonMountPoint self = (CannonMountPoint) (Object) this;
-		if (!(this.getLevel().getBlockEntity(this.pos) instanceof CannonMountBlockEntity mount)) return stack;
-		PitchOrientedContraptionEntity poce = mount.getContraption();
-		if (poce == null || !(poce.getContraption() instanceof AbstractMountedCannonContraption cannon)) return stack;
+		BlockEntity be = this.getLevel().getBlockEntity(this.pos);
+		PitchOrientedContraptionEntity poce;
+		if (be instanceof CannonMountBlockEntity mount) {
+			poce = mount.getContraption();
+		} else if (be instanceof FixedCannonMountBlockEntity mount) {
+			poce = mount.getContraption();
+		} else {
+			return stack;
+		}
+		if (poce == null || !(poce.getContraption() instanceof AbstractMountedCannonContraption cannon))
+			return stack;
 		return self.getInsertedResultAndDoSomething(stack, simulate, cannon, poce);
 	}